refactor!: consolidate Node.js and browser entry points
