+++ tsdown.config.ts
@@ -1,16 +1,7 @@
-import { defineConfig, type UserConfig } from 'tsdown';
+import { defineConfig } from 'tsdown';
 
-const BROWSER_CONFIG: UserConfig = {
-  entry: { "index.browser": "src/index.browser.ts" },
-  format: "esm",
-  platform: "browser",
-};
-
-const NODE_CONFIG: UserConfig = {
-  entry: { "index.node": "src/index.node.ts" },
+export default defineConfig({
+  entry: "src/index.ts",
   format: ["cjs", "esm"],
-  platform: "node",
   copy: ['build/wasm/aptos_confidential_asset_wasm_bg.wasm'],
-};
-
-export default defineConfig([BROWSER_CONFIG, NODE_CONFIG]);
+});
